review changes
test class for each class
inline testList
add non-id compose tests
